+++ app/controllers/frontend/dc20/ancestries_controller.rb
@@ -6,7 +6,13 @@ module Frontend
       include SerializeRelation
 
       def index
-        serialize_relation(::Dc20::Feat.where(origin: 0), ::Dc20::FeatSerializer, :ancestries)
+        serialize_relation_v2(::Dc20::Feat.where(origin: 0), ::Dc20::FeatSerializer, :ancestries, cache_options: cache_options)
+      end
+
+      private
+
+      def cache_options
+        { key: 'dc20_ancestries/v1', expires_in: 24.hours }
       end
     end
   end