+++ app/controllers/frontend/dc20/maneuvers_controller.rb
@@ -6,7 +6,13 @@ module Frontend
       include SerializeRelation
 
       def index
-        serialize_relation(::Dc20::Feat.where(origin: 3), ::Dc20::FeatSerializer, :maneuvers, { except: %i[description price] })
+        serialize_relation_v2(
+          ::Dc20::Feat.where(origin: 3),
+          ::Dc20::FeatSerializer,
+          :maneuvers,
+          cache_options: { key: 'dc20_maneuvers/v1', expires_in: 24.hours },
+          serialized_fields: { except: %i[description price] }
+        )
       end
     end
   end